requirements.txt
pydub
pyannote.audio
soundfile